Transport line side only build logic bugfixes
Bugfixes:
- Fixed itemOffset not updated when placed side only into another segment
- Fixed itemOffset not being adjusted when a segment becomes side only
- Fixed narrowing convertion when calculating the length of a transport segment
- Fixed invalid printf format for debug menu transport segment targetInsertOffset

Changes:
- Refactored logic for updating neighboring transport segments

+++ src/game/logic/transport_line_controller.cpp
@@ -166,7 +166,7 @@ void UpdateSide(const game::TransportLineOffset& tiles_moved, game::TransportSeg
 					// |   |   |   |
 					// 3   2   1   0
 					// targetOffset of 0: Length is 1
-					length = 1 + segment.targetInsertOffset;
+					length = static_cast<double>(1) + segment.targetInsertOffset;
 					break;
 
 				default:
